+++ CHANGELOG.md
@@ -13,6 +13,9 @@ All notable changes to this project will be documented in this file.
     - New filters include access pass type (`prepaid` or `solana-validator`) and Solana identity public key.
     - Updated command arguments and logic, with tests adjusted to cover new options.
 
+- Device controller
+    - Use serviceability onchain delay for link metrics
+
 ## [v0.6.0](https://github.com/malbeclabs/doublezero/compare/client/v0.5.3...client/v0.6.0) – 2025-08-28
 
 ### Breaking